refactor(ts): relocate projectEnemyAfflictionViz to projection/
Move projectEnemyAfflictionViz + its physical-marker normalization helpers out
of the projectOptimizerResult adapter into a typed
simulation/projection/projectEnemyAfflictionViz.ts, alongside the other
projection modules. First step of dissolving the adapters/ layer. No behavior
change (affliction-viz + golden suites unchanged).
